% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{barycentric2cartesian}
\alias{barycentric2cartesian}
\title{Convert vector from barycentric coordinates to Cartesian coordinates}
\usage{
barycentric2cartesian(barycentric, v_cartesian)
}
\arguments{
\item{barycentric}{K-vector representing a vertex in barycentric coordinates.}

\item{v_cartesian}{list of 'reference points' in Cartesian coordinates.}
}
\value{
vertex in Cartesian coordinates.
}
\description{
This function converts a vertex represented in barycentric coordinates,
into a vertex represented in Cartesian coordinates.
}
