% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cartesian2barycentric}
\alias{cartesian2barycentric}
\title{Convert vector from Cartesian coordinates to barycentric coordinates}
\usage{
cartesian2barycentric(cartesian, matrixT, v_cartesian)
}
\arguments{
\item{matrixT}{a matrix}

\item{v_cartesian}{a list of 'reference points' in Cartesian coordinate}

\item{barycentric}{K-1-vector representing a vertex in barycentric coordinates.}
}
\value{
vertex in barycentric coordinates.
}
\description{
This function converts a vertex represented in Cartesian coordinates,
into a vertex represented in barycentric coordinates.
}
