% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dempsterpolytope-package.R
\docType{package}
\name{dempsterpolytope-package}
\alias{dempsterpolytope-package}
\alias{dempsterpolytope}
\title{dempsterpolytope}
\description{
...
}
\details{
...
}
\author{
Pierre E. Jacob <pjacob@g.harvard.edu>
}
\keyword{package}
