% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etas2vertices.R
\name{etas2vertices}
\alias{etas2vertices}
\title{Convert matrix of eta values into polytope in H- and V-representations}
\usage{
etas2vertices(etas)
}
\arguments{
\item{etas}{a matrix of values, such as produced by \code{\link{gibbs_sampler}}.}
}
\value{
A list containing
\enumerate{
\item vertices_barcoord: the vertices of the polytope in barycentric coordinates
\item constr: a list corresponding to the H-representation, with "constr" for the matrix A, "rhs" for the vector b,
and "dir" containing the sign "<=" indicating the direction of all constraints.
}
}
\description{
This function converts a KxK matrix "eta" to a polytope within simplex of dimension K.
The polytope describes a feasible set. 
On barycentric coordinate recall each
constraint is of the form: theta_ell / theta_k <= eta[k,l] i.e. theta_ell -
eta[k,l] theta_k  <= 0 and theta is in the simplex, i.e. sum_{j<K} theta_j <=
1, -theta_j <= 0. The function writes the constraints as a matrix A and
a vector b such that A x <= b then calls a function of the package 'rcdd'
(mimicking what's done in the 'hitandrun' package) to obtain the coordinates of the
vertices of the polytope, i.e. the V-representation.
}
