% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etas_vertices.R
\name{etas_vertices}
\alias{etas_vertices}
\title{Convert matrix of eta values into list of vertices}
\usage{
etas_vertices(etas)
}
\arguments{
\item{etas}{a matrix of values, such as produced by \code{\link{gibbs_sampler}}
or \code{\link{rejection_sampler}}.}
}
\value{
The vertices of the polytope in barycentric coordinates, in an M x K matrix with M rows if there are M vertices.
}
\description{
This function converts a KxK matrix "eta" into
a matrix listing the vertices of the associated polytope within the simplex of dimension K.
The matrix "eta" describes
constraints of the form: theta_l / theta_k <= eta\link{k,l},
or equivalently theta_l -
eta\link{k,l} theta_k  <= 0, on top of the additional constraint that
theta must be in the simplex: i.e. sum_{j<K} theta_j <=
1, -theta_j <= 0.
The function calls the package \code{rcdd}
(mimicking what's done in the \code{hitandrun} package) to obtain the coordinates of the
vertices of the polytope, i.e. the V-representation.
}
\examples{
\dontrun{
etas <- rejection_sampler(c(1,0,3,2))$etas
etas_vertices(etas)
}
}
