% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_us.R
\name{extend_us}
\alias{extend_us}
\title{Extend auxiliary variables to include additional empty categories}
\usage{
extend_us(Us, whichbefore, whichnew)
}
\arguments{
\item{Us}{a list of auxiliary variables.}

\item{whichbefore}{a vector of indices indicating location of previous categories}

\item{whichnew}{a vector of indices indicating location of new (empty) categories}
}
\value{
A list with the following entries:
#' \itemize{
\item "etas": an array of dimension niterations x K x K.
For iteration 'iteration', etas\link{iteration,,} is a K x K matrix
representing a convex polytope.
The feasible set at that iteration, is the set of all vectors
theta such that, theta_l/theta_k < etas\link{iteration,k,l}
for all k,l in {1,...,K}.
\item "Us": a list of K arrays. For iteration 'iteration',
and category 'k' in {1,...,K}, if N_k = 0 then Us[\link{k}] is NA.
If N_k > 0 then Us[\link{k}]\link{iteration,,}
is a matrix with N_k rows, and K columns, representing the auxiliary variables
"U"'s generated at that iteration.
}
}
\description{
Given a list of auxiliary variables Us, such as
generated by \code{\link{gibbs_sampler}},
add empty categories, i.e. zero counts.
}
