% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_pts.R
\name{initialize_pts}
\alias{initialize_pts}
\title{Initialize auxiliary variables given theta}
\usage{
initialize_pts(counts, theta)
}
\arguments{
\item{counts}{a vector of K integers representing counts, possibly including zeros}

\item{theta}{a point in the simplex, represented by a K-vector of non-negative values summing to one}
}
\value{
A list with the following entries.
\itemize{
\item "pts": a list of K matrices of size N_k x K containing the
auxiliary variables, or NA if N_k = 0
\item "minratios", a list of K vectors of size K,
where minratios[\link{k}] has entry ell equal to min_{n in I_k} u_{n,ell} / u_{n,k}
which can be used to define 'eta_{k,ell}'. Equals +Inf if N_k = 0.
}
}
\description{
For a point theta in the simplex, and a vector of counts (N_1,...,N_K),
sample for each k, N_k points uniformly in the subsimplex Delta_k(theta), where N_k > 0,
and compute the minimum ratio u_{n,l} / u_{n,k} for all l in {1,...,K}.
}
\examples{
\dontrun{
initialize_pts(c(3,0,2), c(1/3,1/3,1/3))
}
}
