% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{linkage_cdf_lowerupper}
\alias{linkage_cdf_lowerupper}
\title{## same thing, but ignoring etas[k,.]
get_lower_upper_updatek <- function(etas, A, b, k){
  upper <- 1
  lower <- 0
  K <- dim(etas)[1]
  for (k1 in setdiff(1:K, k)){
    for (k2 in setdiff(1:K, k1)){
      c <- etas[k1,k2]
      denom <- A[k2] - c * A[k1]
      if (denom >= 0){
        upper <- min(upper, (c * b[k1] - b[k2]) / (denom))
      } else {
        lower <- max(lower, (c * b[k1] - b[k2]) / (denom))
      }
    }
  }
  return(c(lower, upper))
}}
\usage{
linkage_cdf_lowerupper(phi_0, lu_chain)
}
\description{
## same thing, but ignoring etas[k,.]
get_lower_upper_updatek <- function(etas, A, b, k){
  upper <- 1
  lower <- 0
  K <- dim(etas)[1]
  for (k1 in setdiff(1:K, k)){
    for (k2 in setdiff(1:K, k1)){
      c <- etas[k1,k2]
      denom <- A[k2] - c * A[k1]
      if (denom >= 0){
        upper <- min(upper, (c * b[k1] - b[k2]) / (denom))
      } else {
        lower <- max(lower, (c * b[k1] - b[k2]) / (denom))
      }
    }
  }
  return(c(lower, upper))
}
}
