% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsolve_over_eta.R
\name{lpsolve_over_eta_log}
\alias{lpsolve_over_eta_log}
\title{Solve LP over polytope described by log(eta)}
\usage{
lpsolve_over_eta_log(eta, objective_vector)
}
\arguments{
\item{eta}{A matrix of dimension K x K describing a convex polytope.}

\item{objective_vector}{a vector giving coefficient of linear function (of log theta) to minimize.}
}
\value{
Evaluation of the objective function at its minimum.
}
\description{
Minimize linear program of the form
t(objective_vector) * log(theta)
over theta in the convex polytope of log(theta)
such that log(theta_l) - log(theta_k) <= log(eta\link{k,l})
and such that sum_k log (theta_k) = 0.
Employs the \code{lpSolveAPI} package.
}
\examples{
\dontrun{
eta <- rejection_sampler(c(1,0,3,2))$etas
lpsolve_over_eta_log(eta, c(1,0,0,0))
}
}
