% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlodsm-package.R
\docType{package}
\name{montecarlodsm-package}
\alias{montecarlodsm-package}
\alias{montecarlodsm}
\title{montecarlodsm}
\description{
...
}
\details{
...
}
\author{
Pierre E. Jacob <pjacob@g.harvard.edu>
}
\keyword{package}
