% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection_sampler.R
\name{rejection_sampler}
\alias{rejection_sampler}
\title{Rejection sampler for DS convex polytopes}
\usage{
rejection_sampler(counts, maxnattempts = 1e+05)
}
\arguments{
\item{counts}{a vector of counts, of length K, that could contain zeros.}

\item{maxnattempts}{an integer indicating the number of trials to perform before giving up.}
}
\value{
A list with the following entries:
\itemize{
\item "etas": a KxK matrix if procedure succeeded, otherwise NULL.
\item "nattempts": an integer indicating the number of trials.
}
}
\description{
Implements a rejection sampler,
as an alternative to the proposed Gibbs sampler. Only works
for very small data sets, otherwise the number of attempts required
to obtain a draw is prohibitively large.
The sampler draws N uniform variables in the simplex of dimension K,
and then checks whether the constraints are satisfied, specifically
by checking whether a certain graph contains negative cycles.
The implementation uses the "igraph" package for that.
}
\examples{
\dontrun{
rejection_sampler(c(1,2,3))
rejection_sampler(c(6,0,5,0))
}
}
