% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejectionsampler.R
\name{rejectionsampler}
\alias{rejectionsampler}
\title{Rejection sampler to obtain convex polytopes for DS inference in Categorical distributions}
\usage{
rejectionsampler(counts)
}
\arguments{
\item{counts}{a vector of counts, of length K
It returns accepted states only, and 
and the number of attempts it took}
}
\description{
This is implemented to perform quick comparison on very small data sets; the method does not 
scale well with the number of counts. It samples auxiliary variables uniformly in the simplex,
and then checks whether constraints are satisfied. The implementation uses the igraph package.
}
\examples{
\dontrun{
rejectionsampler(c(1,2,3,1,2,3,2,2,1), 3)
}
}
