% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_upper_bound.R
\name{tv_upper_bound}
\alias{tv_upper_bound}
\title{Turns meeting times into upper bounds on TV to stationarity}
\usage{
tv_upper_bound(meetingtimes, L, t)
}
\arguments{
\item{meetingtimes}{vector of meeting times, such as generated by \code{\link{sample_meeting_times}}.}

\item{t}{iteration at which to compute the TV upper bound.}

\item{lag}{lag between the chains.}
}
\value{
A non-negative value representing an
estimated upper bound between the chain at iteration t and the limiting distribution
of the chain.
}
\description{
Given a vector of meeting times, e.g. generated by \code{\link{sample_meeting_times}}, a choice of lag L
and an iteration t, returns
mean(pmax(0,ceiling((meetingtimes-L-t)/L))).
}
\examples{
\dontrun{
nrep <- 100
meeting_times <- sapply(1:nrep, function(irep) sample_meeting_times(counts = c(3,2,0,1)))
tmax <- floor(max(meeting_times)*1.2)
ubounds <- sapply(1:tmax, function(t) tv_upper_bound(meeting_times, 1, t))
plot(x = 1:tmax, y = ubounds, type = 'l', xlab = "iteration", ylab = "TV upper bounds")
}
}
