% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dempsterpolytope-package.R
\docType{package}
\name{dempsterpolytope-package}
\alias{dempsterpolytope-package}
\alias{dempsterpolytope}
\title{Gibbs sampler for DS inference in Categorical distributions}
\description{
Implements a Gibbs sampler for Dempster's approach to inference in Categorical distributions;
}
\details{
This package is a companion to an article [...] the goal of the package is merely to reproduce the experiments.
}
\author{
anonymous
}
\keyword{package}
