% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dempsterpolytope-package.R
\docType{package}
\name{dempsterpolytope-package}
\alias{dempsterpolytope-package}
\alias{dempsterpolytope}
\title{Gibbs sampler for Dempster's inference in Categorical distributions}
\description{
Implements a Gibbs sampler for Dempster's inference approach for Categorical distributions; 
this package is a companion to an article by Pierre E. Jacob, Ruobin Gong, Paul T. Edlefsen, Arthur P. Dempster
}
\details{
The algorithm is described in an upcoming article. The goal of the package is merely to reproduce the experiments.
}
\author{
Pierre E. Jacob <pierre.jacob.work@gmail.com>
}
\keyword{package}
