% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbssampler.R
\name{initialize_pts}
\alias{initialize_pts}
\title{Initialize auxiliary variables}
\usage{
initialize_pts(counts, theta)
}
\arguments{
\item{counts}{a vector of K integers representing counts}

\item{theta}{a point in the simplex, represented by a K-vector of non-negative values summing to one}
}
\value{
A list with 'pts', containing a list of K matrices of size N_k x K containing the auxiliary variables
and 'minratios', containing a list of K vectors of size K, where minratios[[k]] has entry ell equal to min_{n in I_k} u_{n,ell} / u_{n,k}
which are later used to define 'eta'. The notation follows that of the paper.
}
\description{
For a point theta in the simplex, a vector of counts, sample points 'u_n' uniformly in the subsimplex Delta_k(theta) 
where x_n = k.
}
\examples{
initialize_pts(c(3,1,2), c(1/3,1/3,1/3))
}
